requirements.txt
asgiref==3.5.2
Django==4.1.2
psycopg2==2.9.6
sqlparse==0.4.3
django-dbsettings==1.3.0
djangorestframework==3.14.0